+++ cassandra-operator/tools.go
@@ -10,6 +10,6 @@ import (
 	_ "github.com/sky-uk/licence-compliance-checker"
 	_ "golang.org/x/lint/golint"
 	_ "golang.org/x/tools/cmd/goimports"
-	_ "sigs.k8s.io/controller-tools/cmd/controller-gen"
 	_ "k8s.io/code-generator"
+	_ "sigs.k8s.io/controller-tools/cmd/controller-gen"
 )